# examples/petclinic/internal/spicegen/mysql/artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mysql API",
    "version": "0.1.0-dev"
  },
  "paths": {
...
    "/owners": {
      "get": {
        "operationId": "Controller_Find",
        "summary": "Find",
        "parameters": [
          {
            "name": "lastName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "See Other"
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "x-spice-symbol": "spice:symbol:v1|method|57:github.com/spice-framework/spice/examples/petclinic/owner|10:Controller|4:Find"
      }
    },
    "/owners/find": {
      "get": {
        "operationId": "Controller_FindForm",
        "summary": "FindForm",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "See Other"
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "x-spice-symbol": "spice:symbol:v1|method|57:github.com/spice-framework/spice/examples/petclinic/owner|10:Controller|8:FindForm"
      }
    },
    "/owners/new": {
      "get": {
        "operationId": "Controller_NewForm",
        "summary": "NewForm",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "See Other"
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "x-spice-symbol": "spice:symbol:v1|method|57:github.com/spice-framework/spice/examples/petclinic/owner|10:Controller|7:NewForm"
      },
      "post": {
        "operationId": "Controller_Create",
        "summary": "Create",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "telephone": {
                    "type": "string"
                  }
                },
                "required": [
                  "address",
                  "city",
                  "firstName",
                  "lastName",
                  "telephone"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "See Other"
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "x-spice-symbol": "spice:symbol:v1|method|57:github.com/spice-framework/spice/examples/petclinic/owner|10:Controller|6:Create"
      }
    },
    "/owners/{ownerId}": {
      "get": {
        "operationId": "Controller_Show",
        "summary": "Show",
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "See Other"
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "x-spice-symbol": "spice:symbol:v1|method|57:github.com/spice-framework/spice/examples/petclinic/owner|10:Controller|4:Show"
      }
    },
    "/owners/{ownerId}/edit": {
      "get": {
        "operationId": "Controller_EditForm",
        "summary": "EditForm",
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "See Other"
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "x-spice-symbol": "spice:symbol:v1|method|57:github.com/spice-framework/spice/examples/petclinic/owner|10:Controller|8:EditForm"
      },
      "post": {
        "operationId": "Controller_Update",
        "summary": "Update",
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "telephone": {
                    "type": "string"
                  }
                },
                "required": [
                  "address",
                  "city",
                  "firstName",
                  "lastName",
                  "telephone"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "See Other"
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "x-spice-symbol": "spice:symbol:v1|method|57:github.com/spice-framework/spice/examples/petclinic/owner|10:Controller|6:Update"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SpiceProblem": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "format": "uri-reference"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
    }
  }
}